requirements.txt
# Placeholder for requirements.txt
Flask
requests
python-whois
